[REFACTOR] exception 폴더 변경
